requirements.txt
pyautogui
opencv-python
numpy
pillow
pywinauto
pytesseract
apscheduler
python-dotenv
pyyaml
loguru
